satellite/requirements-satellite.txt
grpcio
grpcio-tools
PyAstronomy~=0.16.0
numpy~=1.20.3
networkx~=2.5.1
scipy==1.6.3
matplotlib==3.4.2
flask==2.0.0
requests==2.25.1
h3~=3.7.2
toml==0.10.2